Interpret view name as the file name
This is different from how it works today.

Now you have to declare a name and a file name, but in the future you just need to give the filename and the filename itself will be the name of the view. This makes it easy to write scripts that generate the "views.config" file.